Split motor files into separate files and introduced namespace Motors (#522)
* Renamed some files to match the class name.

* Split out motor class into different files.

* Fixed last newline in cpp/h files
