+++ Grbl_Esp32/src/Motors/NullMotor.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "Motor.h"
+
+namespace Motors {
+    class Nullmotor : public Motor {};
+}
