+++ Grbl_Esp32/src/Motors/RcServoSettings.h
@@ -1,24 +1,5 @@
 #pragma once
 
-/*
-    RcServoClass.h
-    
-    Part of Grbl_ESP32
-
-    2020 -	Bart Dring
-    
-    Grbl is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-    Grbl is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-    You should have received a copy of the GNU General Public License
-    along with Grbl.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
 // this is the pulse range of a the servo. Typical servos are 0.001 to 0.002 seconds
 // some servos have a wider range. You can adjust this here or in the calibration feature
 #define SERVO_MIN_PULSE_SEC 0.001  // min pulse in seconds
